chore: update dependencies and fix related issues for upgrade
Former-commit-id: eb40e428c4fa93f2f8b9f4388c3b5cf46d315098
Former-commit-id: 5ae939080aa8c47492f58837a4a8fd7fce5c28d8
Former-commit-id: 10cc09928a02eb976ec9d8b3be66301878b1d6a0

+++ src/graph.cpp
@@ -104,7 +104,8 @@ auto Graph::MarkConnectedComponents() -> std::vector<ComponentInfo> {
 #ifndef NDEBUG
   static const auto isUnassigned = [](NodeContainer::const_reference p) { return p.second->ComponentID == 0; };
 #endif
-  LANCET_ASSERT(std::count_if(nodesMap.cbegin(), nodesMap.cend(), isUnassigned) == nodesMap.size());  // NOLINT
+  LANCET_ASSERT(static_cast<std::size_t>(std::count_if(nodesMap.cbegin(), nodesMap.cend(), isUnassigned)) ==
+                nodesMap.size());  // NOLINT
 
   for (NodeContainer::reference p : nodesMap) {
     if (p.second->ComponentID != 0) continue;
